+++ src/modules/tools/wfsSearch/utils/requests.js
@@ -4,6 +4,8 @@ import {Point, Polygon} from "ol/geom";
 import WFS from "ol/format/WFS";
 import handleAxiosResponse from "../../../../utils/handleAxiosResponse";
 import {buildFilter, buildStoredFilter} from "./buildFilter";
+import store from "../../../../app-store";
+import crs from "@masterportal/masterportalapi/src/crs";
 
 /**
  * Makes sure that the filter is ready to be used.
@@ -30,13 +32,15 @@ function adjustFilter (filter) {
  * @param {string} responseData The response data returned by the WFS-G; GML string.
  * @param {String|String[]} namespaces The namespaces of the service.
  * @param {String} memberSuffix The suffix of the feature in the FeatureCollection.
+ * @param {String} responseProjection The expected projection of the features in responseData
  * @returns {module:ol/Feature[]} Array of Features returned from the service.
  */
-function parseGazetteerResponse (responseData, namespaces, memberSuffix) {
+function parseGazetteerResponse (responseData, namespaces, memberSuffix, responseProjection) {
     const attributes = {},
         features = [],
         namespaceArray = Array.isArray(namespaces) ? namespaces : [namespaces],
-        gmlFeatures = new DOMParser().parseFromString(responseData, "application/xml").getElementsByTagName(`wfs:${memberSuffix}`);
+        gmlFeatures = new DOMParser().parseFromString(responseData, "application/xml").getElementsByTagName(`wfs:${memberSuffix}`),
+        transformCoordinates = responseProjection !== store.getters["Maps/projection"].getCode();
 
     if (gmlFeatures.length === 0) {
         return features;
@@ -53,16 +57,18 @@ function parseGazetteerResponse (responseData, namespaces, memberSuffix) {
             });
 
         if (feature.getElementsByTagName("iso19112:position").length > 0) {
-            attributes.geometry = new Point(
-                feature.getElementsByTagName("iso19112:position")[0].getElementsByTagName("gml:pos")[0]
-                    .textContent.split(" ")
-                    .map(coordinate => parseFloat(coordinate))
-            );
+            let coordinates = feature.getElementsByTagName("iso19112:position")[0].getElementsByTagName("gml:pos")[0]
+                .textContent.split(" ")
+                .map(coordinatePart => parseFloat(coordinatePart));
+
+            if (transformCoordinates) {
+                coordinates = crs.transformToMapProjection(mapCollection.getMap(store.getters["Maps/mode"]), responseProjection, coordinates);
+            }
+            attributes.geometry = new Point(coordinates);
         }
 
         if (feature.getElementsByTagName("iso19112:geographicExtent").length > 0) {
-            // Reads the coordinates from the response and prepares them to a readable format for the OL Polygon.
-            attributes.geographicExtent = new Polygon([
+            let coordinates =
                 feature
                     .getElementsByTagName("iso19112:geographicExtent")[0]
                     .getElementsByTagName("gml:posList")[0]
@@ -70,8 +76,15 @@ function parseGazetteerResponse (responseData, namespaces, memberSuffix) {
                     .map(coordinate => parseFloat(coordinate))
                     .reduce((accumulator, _, index, array) => index % 2 === 0
                         ? [...accumulator, [array[index], array[index + 1]]]
-                        : accumulator, [])
-            ]);
+                        : accumulator, []);
+
+            if (transformCoordinates) {
+                coordinates = coordinates.map(coordinate => {
+                    return crs.transformToMapProjection(mapCollection.getMap(store.getters["Maps/mode"]), responseProjection, coordinate);
+                });
+            }
+            // Reads the coordinates from the response and prepares them to a readable format for the OL Polygon.
+            attributes.geographicExtent = new Polygon(coordinates);
         }
 
         features.push(new Feature(attributes));
@@ -106,7 +119,6 @@ let currentRequest = null;
 /**
  * Searches the service for features based on the build or given filter.
  *
- * @param {Object} store Vuex store.
  * @param {Object} currentInstance The currently selected searchInstance.
  * @param {Object[]} currentInstance.literals Array of literals.
  * @param {?Object} [currentInstance.requestConfig.gazetteer] Declares whether the used WFS service is a WFS-G, which needs to be parsed differently.
@@ -115,12 +127,13 @@ let currentRequest = null;
  * @param {String} currentInstance.requestConfig.layerId Id of the layer defined in the services.json. Here it is used to check if that is the case or if the layer was defined in the rest-service.json.
  * @param {Number} currentInstance.requestConfig.maxFeatures The maximum amount of features allowed.
  * @param {String} currentInstance.requestConfig.storedQueryId Id of the stored Query. If given, a WFS@2.0.0 is queried.
+ * @param {String} currentInstance.requestConfig.responseProjection The expected projection of the features in responseData.
  * @param {Object} service The service to send the request to.
  * @param {?String} [singleValueFilter = null] If given, this filter should be used.
  * @param {?String} [featureType = null] FeatureType of the features which should be requested. Only given for queries for suggestions.
  * @returns {Promise} If the send request was successful, the found features are converted from XML to OL Features and returned.
  */
-export function searchFeatures (store, {literals, requestConfig: {gazetteer = null, layerId, maxFeatures, storedQueryId}}, service, singleValueFilter = null, featureType = null) {
+export function searchFeatures ({literals, requestConfig: {gazetteer = null, layerId, maxFeatures, storedQueryId, responseProjection = store.getters["Maps/projection"].getCode()}}, service, singleValueFilter = null, featureType = null) {
     const fromServicesJson = Boolean(layerId);
     let filter;
 
@@ -131,11 +144,11 @@ export function searchFeatures (store, {literals, requestConfig: {gazetteer = nu
         filter = storedQueryId ? buildStoredFilter(literals) : buildFilter(literals);
     }
 
-    return sendRequest(store, service, filter, fromServicesJson, storedQueryId, maxFeatures, featureType)
+    return sendRequest(service, filter, fromServicesJson, storedQueryId, maxFeatures, featureType)
         .then(data => {
             // NOTE: This extra case can be removed when OL can parse WFS-G services with the WFS.readFeatures function.
             if (gazetteer) {
-                return parseGazetteerResponse(data, gazetteer.namespaces, gazetteer.memberSuffix);
+                return parseGazetteerResponse(data, gazetteer.namespaces, gazetteer.memberSuffix, responseProjection);
             }
             return new WFS({version: storedQueryId ? "2.0.0" : "1.1.0"}).readFeatures(data);
         });
@@ -144,7 +157,6 @@ export function searchFeatures (store, {literals, requestConfig: {gazetteer = nu
 /**
  * Builds the request url and sends a GetFeature request via GET to the service.
  *
- * @param {Object} store Vuex store.
  * @param {Object} service The service to send the request to.
  * @param {String} service.url The base Url as defined in the services.json or rest-services.json.
  * @param {String} service.typeName If the Url was defined in the services.json, the typeName is set to be added to the Url.
@@ -156,7 +168,7 @@ export function searchFeatures (store, {literals, requestConfig: {gazetteer = nu
  * @returns {Promise} If the request was successful, the data of the response gets resolved.
  *                    If an error occurs (e.g. the service is not reachable or there was no such feature) the error is caught and the message is displayed as an alert.
  */
-function sendRequest (store, {url, typeName}, filter, fromServicesJson, storedQueryId, maxFeatures = 8, featureType = null) {
+function sendRequest ({url, typeName}, filter, fromServicesJson, storedQueryId, maxFeatures = 8, featureType = null) {
     let requestUrl = url;
 
     if (fromServicesJson) {